:hammer: Refactor: 커뮤니티 리팩토링
